renamed files to be more indicative of what the function is
