+++ query/component_geocoding.js
@@ -102,7 +102,7 @@ function generateQuery( clean ){
 
   var q = getQuery(vs);
 
-  console.log(JSON.stringify(q.body, null, 2));
+  // console.log(JSON.stringify(q.body, null, 2));
 
   return q;
 }
