+++ test/unit/query/component_geocoding.js
